feat: name, prompt_2, try, except
